backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing
Pillow==10.1.0
exifread==3.0.0

# OpenAI for GPT-4o Vision
openai==1.3.5

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS support
python-jose[cryptography]==3.3.0

# File handling
aiofiles==23.2.1

# Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Data processing
numpy==1.26.2
pandas==2.1.3